requirements.txt
pyyaml>=6.0.1
requests>=2.31.0
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
matplotlib>=3.5.0
seaborn>=0.11.0
gql>=3.4.0
mcp>=0.1.0
openai>=1.0.0
aiohttp>=3.8.0
xgboost>=1.7.0
python-dotenv>=1.0.0
